Move requisition item info
